Android X migration - included copy of the original source
Android X is the new environment that Android Apps should use as there is no longer support for AppCompat libraries, so it is important to migrate the application to Android X as there will be backwards compatibility and even more performance optimizations
