+++ app/src/androidTest/java/sirtapp/sirtdetection/com/sirtapp/ExampleInstrumentedTest.java
@@ -1,8 +1,8 @@
 package sirtapp.sirtdetection.com.sirtapp;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.InstrumentationRegistry;
+import androidx.test.runner.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
