--- app/src/main/java/sirtapp/sirtdetection/com/sirtapp/Token.java
@@ -1,21 +1,0 @@
-package sirtapp.sirtdetection.com.sirtapp;
-
-public class Token {
-    private int status;
-    private String token;
-
-    public String getToken() {
-        return token;
-    }
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-    public void setStatus(int status) {
-        this.status = status;
-    }
-}
-
