+++ build.gradle
@@ -4,6 +4,9 @@ buildscript {
     
     repositories {
         google()
+        maven {
+            url  "http://dl.bintray.com/piasy/maven"
+        }
         jcenter()
     }
     dependencies {
@@ -18,6 +21,9 @@ buildscript {
 allprojects {
     repositories {
         google()
+        maven {
+            url  "http://dl.bintray.com/piasy/maven"
+        }
         jcenter()
     }
 }
