--- image-zoom-view/src/main/java/ozaydin/serkan/com/image_zoom_view/SaveFileListener.java
@@ -1,10 +1,0 @@
-package ozaydin.serkan.com.image_zoom_view;
-
-import java.io.File;
-
-public interface SaveFileListener {
-
-void onSuccess(File file);
-void onFail(Exception exception);
-
-}
