--- image-zoom-view/src/main/res/anim/scale_out.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android">
-
-
-    <scale
-        android:duration="@android:integer/config_shortAnimTime"
-        android:fromXScale="1.0"
-        android:fromYScale="1.0"
-        android:toXScale="0"
-        android:toYScale="0"
-        android:pivotX="50%"
-        android:pivotY="50%"
-        />
-
-    <alpha android:fromAlpha="1.0"
-        android:toAlpha="0"
-        android:duration="@android:integer/config_shortAnimTime"
-        android:interpolator="@android:anim/decelerate_interpolator"
-        />
-</set>
