--- image-zoom-view/src/main/res/drawable/ic_file_download_black_24dp.xml
@@ -1,5 +1,0 @@
-<vector android:height="24dp" android:tint="#0C0C0C"
-    android:viewportHeight="24.0" android:viewportWidth="24.0"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M19,9h-4V3H9v6H5l7,7 7,-7zM5,18v2h14v-2H5z"/>
-</vector>
