--- image-zoom-view/src/main/res/layout/bottom_sheet_layout.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    >
-
-    <LinearLayout
-        android:visibility="gone"
-        android:id="@+id/bottom_sheet_layout_download_image_linear_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-        <ImageView
-            android:src="@drawable/ic_file_download_black_24dp"
-            android:layout_width="40dp"
-            android:layout_height="40dp" />
-        <TextView
-            android:textSize="18sp"
-            android:textStyle="bold"
-            android:text="@string/description_save"
-            android:layout_marginLeft="5dp"
-            android:layout_gravity="center"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/bottom_sheet_layout_download_image_text_view"/>
-
-    </LinearLayout>
-
-
-</LinearLayout>
