--- image-zoom-view/src/main/res/layout/image_view_zoom_bottom_sheet_layout.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-
-
-
-
-
-
-</LinearLayout>
