--- image-zoom-view/src/main/res/values/attrs.xml
@@ -1,5 +1,0 @@
-<resources>
-    <declare-styleable name="ImageViewZoom">
-        <attr name="circle" format="boolean" />
-    </declare-styleable>
-</resources>
