--- image-zoom-view/src/main/res/values/color.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <color name="black">#000</color>
-    <color name="transprent_black">#8000</color>
-</resources>
