--- image-zoom-view/src/main/res/values/styles.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-
-    <style name="DialogStyle">
-
-        <item name="android:windowEnterAnimation"> @anim/scale_in</item>
-        <item name="android:windowExitAnimation">@anim/scale_out</item>
-    </style>
-</resources>
